fix safe-transmute handling of enums
